fix: implementation of weapon and armor

+++ server/src/GameLogic/Bullets.cs
@@ -1,6 +1,5 @@
 namespace Thuai.Server.GameLogic;
 
-
 /// <summary>
 /// Bullet used by a weapon.
 /// </summary>
@@ -19,7 +18,7 @@ public interface IBullet
 
     public Position BulletPosition { get; set; }
     public double BulletSpeed { get; }
-    public double BulletDamage { get; }
+    public int BulletDamage { get; }
 
     public bool AntiArmor { get; }
 
@@ -29,26 +28,26 @@ public interface IBullet
 /// <summary>
 /// Default bullet for a weapon. Used by Cannon.
 /// </summary>
-public class Bullet(Position position, double speed, double damage, bool antiArmor = false) : IBullet
+public class Bullet(Position position, double speed, int damage, bool antiArmor = false) : IBullet
 {
     public IBullet.BulletType Type => IBullet.BulletType.Bullet;
 
     public Position BulletPosition { get; set; } = position;
 
     public double BulletSpeed { get; } = speed;
-    public double BulletDamage { get; } = damage;
+    public int BulletDamage { get; } = damage;
 
     public bool AntiArmor { get; } = antiArmor;
 }
 
-public class LaserBullet(Position position, double speed, double damage, bool antiArmor = false) : IBullet
+public class LaserBullet(Position position, double speed, int damage, bool antiArmor = false) : IBullet
 {
     public IBullet.BulletType Type => IBullet.BulletType.Bullet;
 
     public Position BulletPosition { get; set; } = position;
 
     public double BulletSpeed { get; } = speed;
-    public double BulletDamage { get; } = damage;
+    public int BulletDamage { get; } = damage;
 
     public bool AntiArmor { get; } = antiArmor;
 }
